refactor(Player): Name 객체 Player로 통합
