+++ src/main/java/domain/player/Player.java
@@ -4,15 +4,38 @@ import java.util.Objects;
 
 public class Player {
 
-    private final Name name;
+    private static final String NAME_LENGTH_ERROR_MESSAGE = "참가자 이름의 길이는 1이상 5이하만 가능합니다.";
+    private static final String NAME_FORMAT_ERROR_MESSAGE = "참가자 이름에는 공백이 들어갈 수 없습니다.";
+    private static final int MIN_NAME_LENGTH = 1;
+    private static final int MAX_NAME_LENGTH = 5;
+    private static final String SPACE = " ";
+
+    private final String name;
 
     public Player(String name) {
-        this.name = new Name(name);
+        validate(name);
+        this.name = name;
     }
 
+    private void validate(String name) {
+        validateLength(name);
+        validateSpace(name);
+    }
+
+    private void validateLength(String name) {
+        if (name.length() < MIN_NAME_LENGTH || name.length() > MAX_NAME_LENGTH) {
+            throw new IllegalArgumentException(NAME_LENGTH_ERROR_MESSAGE);
+        }
+    }
+
+    private void validateSpace(String name) {
+        if (name.contains(SPACE)) {
+            throw new IllegalArgumentException(NAME_FORMAT_ERROR_MESSAGE);
+        }
+    }
 
     public String getName() {
-        return name.getName();
+        return name;
     }
 
     @Override
@@ -23,12 +46,12 @@ public class Player {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Player player = (Player) o;
-        return Objects.equals(name.getName(), player.name.getName());
+        Player name1 = (Player) o;
+        return Objects.equals(getName(), name1.getName());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(name.getName());
+        return Objects.hash(getName());
     }
 }
