+++ src/test/java/domain/player/PlayerTest.java
@@ -4,19 +4,34 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatNoException;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-
-public class PlayerTest {
+class PlayerTest {
 
     @ParameterizedTest
     @ValueSource(strings = {"gray", "encho", "pobi", "ann"})
-    @DisplayName("이름의 길이가 1~5인 경우, Player 객체가 생성된다.")
+    @DisplayName("이름의 길이가 1~5인 경우, 예외가 발생하지 않는다.")
     void validateNameLength_Success(String name) {
-        Player player = new Player(name);
-
-        assertThat(player.getName()).isEqualTo(name);
+        assertThatNoException().isThrownBy(() -> new Player(name));
     }
 
+    @ParameterizedTest
+    @ValueSource(strings = {"gray1234", ""})
+    @DisplayName("이름의 길이가 1~5가 아닌 경우, 예외가 발생한다.")
+    void validateNameLength_Fail(String name) {
+        assertThatThrownBy(() -> new Player(name))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("참가자 이름의 길이는 1이상 5이하만 가능합니다.");
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {" gray", "ann ", "po bi", " "})
+    @DisplayName("이름에 공백이 있는 경우 예외가 발생한다.")
+    void validateNameWithSpace_Fail(String name) {
+        assertThatThrownBy(() -> new Player(name))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("참가자 이름에는 공백이 들어갈 수 없습니다.");
+    }
 
 }
